公共方法的抽离
